backend/pyproject.toml
[project]
name = "scrittura-libro-backend"
version = "0.1.0"
description = "Backend per agente di scrittura romanzi"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
    "reportlab>=4.0.0",
    "google-generativeai>=0.8.0",
    "google-genai>=1.0.0",
    "xhtml2pdf>=0.2.11",
    "markdown>=3.4.0",
    "python-multipart>=0.0.21",
    "ebooklib>=0.18",
    "python-docx>=1.1.0",
    "motor>=3.3.0",
    "pymongo>=4.6.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "itsdangerous>=2.1.2",
    "python-multipart>=0.0.21",
    "aiofiles>=23.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.scripts]
start = "uvicorn.app:main"

